eks/agents/a2a/multi-agent/pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"

dependencies = [
    "strands-agents[a2a]>=1.9.1",
    "strands-agents-tools>=0.2.8",
    "fastapi>=0.118.0",
]

[project.scripts]
"multi-agent"                   = "src.travel_agent:main"
"multi-agent-restapi-server"    = "src.travel_agent:run_restapi_server"
"multi-agent-a2a-server"        = "src.travel_agent:run_a2a_server"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
